requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1751547525079/work
accelerate==0.30.0
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1636085263707/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi==2025.7.14
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046050867/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1649184137891/work
charset-normalizer==3.4.2
click @ file:///home/conda/feedstock_root/build_artifacts/click_1758270398201/work
cmake==4.1.0
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
diffusers==0.23.1
filelock==3.18.0
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1739279199749/work
fonttools==4.59.0
fsspec==2025.7.0
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1749108257037/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1688235217226/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
grpcio @ file:///croot/grpc-suite_1681912592597/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1637963920366/work
hf-xet==1.1.5
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
huggingface-hub==0.19.4
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2==3.1.4
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1679712543761/work/keras-2.12.0-py2.py3-none-any.whl
kiwisolver==1.4.8
lit==18.1.8
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1757093412127/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1648737547956/work
matplotlib==3.10.3
mpmath==1.3.0
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1648882415384/work
mypy_extensions==1.1.0
networkx==3.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1750415243897/work
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==1.5.3
peft==0.10.0
pillow==11.3.0
protobuf==3.20.3
psutil==7.0.0
pyasn1 @ file:///home/conda/feedstock_root/build_artifacts/pyasn1_1733217608156/work
pyasn1_modules @ file:///home/conda/feedstock_root/build_artifacts/pyasn1-modules_1743436035994/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1732782409051/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
pyparsing==3.2.3
pyre-extensions==0.0.29
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytz==2025.2
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1733738580568/work
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1733772243268/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1744825426113/work
safetensors==0.5.3
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073872310/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sympy==1.13.3
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1683119049001/work/tensorboard-2.12.3-py3-none-any.whl#sha256=b4a69366784bc347e02fbe7d847e01896a649ca52f8948a11005e205dcf724fb
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=4428d660733184391dab7b6b6a953467b614b984df0cc1baf3d8c5d3d4ecf29d
tensorflow @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-base_1682449491990/work/tensorflow_pkg/tensorflow-2.12.0-cp310-cp310-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-estimator_1682445976941/work/tensorflow_estimator-2.12.0-py2.py3-none-any.whl
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1746039537955/work
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.30.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1651495243689/work
xformers==0.0.20
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1648966516552/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard @ file:///croot/zstandard_1731356346222/work
